+++ .mdlrc.rb
@@ -1,0 +1,4 @@
+all
+rule 'MD004', :style => :sublist
+rule 'MD007', :indent => 4
+rule 'MD029', :style => :ordered